+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+### 4.4 / 2411xx
+
+* new shortcode: listmarker
+* error handling multielevation
+* WordPress 6.7
+
 ### 4.3.5 / 241012
 
 * more WordPress Coding Standards